requirements.txt
ccxt>=4.4.0
python-telegram-bot>=21.0
python-dotenv
colorama
pandas
numpy
requests
